backend/pyproject.toml
[tool.poetry]
name = "nacionalsign-backend"
version = "0.1.0"
description = "Backend service for NacionalSign digital signature SaaS"
authors = ["NacionalSign Team <dev@nacionalsign.io>"]
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = { extras = ["standard"], version = "^0.30.0" }
sqlmodel = "^0.0.16"
sqlalchemy = "^2.0.30"
psycopg2-binary = "^2.9.9"
redis = "^5.0.4"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
python-multipart = "^0.0.9"
httpx = "^0.27.0"
boto3 = "^1.34.0"
pydantic-settings = "^2.3.0"
cryptography = "^42.0.0"
pyotp = "^2.9.0"
bcrypt = "==4.0.1"
twilio = "^9.0.0"
jinja2 = "^3.1.4"
reportlab = "^4.2.2"
pyhanko = "^0.20.0"
python-docx = "^1.1.0"
Pillow = "^10.3.0"
pypdf = "^5.1.0"
psycopg = {version = "^3.1", extras = ["binary"]}

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.0"
pytest-asyncio = "^0.23.6"
httpx = "^0.27.0"
ruff = "^0.4.8"

[build-system]
requires = ["poetry-core>=1.5.0"]
build-backend = "poetry.core.masonry.api"
